+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Braintree iOS SDK Release Notes
 
+## 6.29.0 (2025-02-24)
+* BraintreePayPal
+  * Add PayPal App Switch checkout flow (BETA)
+    * Add `BTPayPalCheckoutRequest(userAuthenticationEmail:enablePayPalAppSwitch:amount:intent:userAction:offerPayLater:currencyCode:requestBillingAgreement:)`
+    * **Note:** This feature is currently in beta and may change or be removed in future releases.
+* BraintreeApplePay
+  * Add `BTApplePayCardNonce.isDeviceToken` for MPAN identification
+
 ## 6.28.0 (2025-02-05)
 * BraintreeVenmo
   * Allow universal links to be set without a return URL scheme (fixes #1505)